Move File Upload Logic to it's own service
